Recursively split buckets.
Previously, all buckets had the same size, so when splitting we were assured
that we would have space for all nodes.  Now that buckets have variable size,
we need to split buckets recursively in order to avoid dropping nodes.

This implementation is iterative, since I find it easier to reason about
this kind of code.